+++ modules/Material/Dialog.qml
@@ -1,6 +1,6 @@
 /*
  * QML Material - An application framework implementing Material Design.
- * Copyright (C) 2014 Michael Spencer
+ * Copyright (C) 2014-2015 Michael Spencer <sonrisesoftware@gmail.com>
  * Copyright (C) 2015 Bogdan Cuza <bogdan.cuza@hotmail.com>
  * Copyright (C) 2015 Mikhail Ivchenko <ematirov@gmail.com>
  *
@@ -23,24 +23,33 @@ import QtQuick.Layouts 1.1
 import Material 0.1
 import Material.Extras 0.1
 
-Popover {
+PopupBase {
     id: dialog
 
-    anchor: Item.Center
-
     overlayLayer: "dialogOverlayLayer"
+    overlayColor: Qt.rgba(0, 0, 0, 0.3)
 
-    backdropColor: Qt.rgba(0, 0, 0, 0.3)
+    opacity: showing ? 1 : 0
+    visible: opacity > 0
 
-    width: Math.max(dialogContainer.width, minimumWidth)
-    height: dialogContainer.height
+    width: Math.max(minimumWidth,
+                    (placeholder.childrenRect.width +
+                     placeholder.anchors.leftMargin +
+                     placeholder.anchors.rightMargin))
+
+    height: titleLabel.height +
+            titleLabel.anchors.topMargin +
+            placeholder.childrenRect.height +
+            placeholder.anchors.topMargin +
+            placeholder.anchors.bottomMargin +
+            buttonContainer.height
 
     property int minimumWidth: units.dp(240)
 
     property alias title: titleLabel.text
 
-    property string negativeButtonText: "CANCEL"
-    property string positiveButtonText: "OK"
+    property string negativeButtonText: "Cancel"
+    property string positiveButtonText: "Ok"
 
     property bool hasActions: true
 
@@ -49,50 +58,34 @@ Popover {
     signal accepted()
     signal rejected()
 
-    transitions: [
-        Transition {
-            ParallelAnimation {
-                NumberAnimation {
-                    target: internalView
-                    property: "opacity"
-                    duration: 200
-                    easing.type: Easing.InOutQuad
-                }
-
-                NumberAnimation {
-                    target: internalView
-                    property: "width"
-                    duration: 100
-                    easing.type: Easing.InOutQuad
-                }
-
-                NumberAnimation {
-                    target: internalView
-                    property: "height"
-                    duration: 100
-                    easing.type: Easing.InOutQuad
-                }
-            }
+    anchors {
+        centerIn: parent
+        verticalCenterOffset: showing ? 0 : -(dialog.height/3)
+        Behavior on verticalCenterOffset {
+            NumberAnimation { duration: 200 }
         }
-    ]
+    }
 
-    Item {
+    Behavior on opacity {
+        NumberAnimation { duration: 200 }
+    }
+
+    Keys.onEscapePressed: dialog.close()
+
+    function show() {
+        open()
+    }
+
+    View {
         id: dialogContainer
-        width: Math.max(minimumWidth,
-                        (placeholder.childrenRect.width +
-                        placeholder.anchors.leftMargin +
-                        placeholder.anchors.rightMargin))
 
-        height: titleLabel.height +
-                titleLabel.anchors.topMargin +
-                placeholder.childrenRect.height +
-                placeholder.anchors.topMargin +
-                placeholder.anchors.bottomMargin +
-                buttonContainer.height
+        anchors.fill: parent
+        elevation: 5
 
         MouseArea {
             anchors.fill: parent
             propagateComposedEvents: false
+
             onClicked: {
                 mouse.accepted = false
             }
@@ -100,8 +93,9 @@ Popover {
 
         Label {
             id: titleLabel
-            height: implicitHeight
+
             style: "title"
+
             anchors {
                 top: parent.top
                 topMargin: units.dp(16)
@@ -111,17 +105,19 @@ Popover {
         }
 
         Item {
+            id: placeholder
+
             anchors {
                 left: parent.left
-                leftMargin: units.dp(16)
                 right: parent.right
-                rightMargin: units.dp(16)
                 top: titleLabel.bottom
-                topMargin: units.dp(16)
                 bottom: buttonContainer.top
+
+                leftMargin: units.dp(16)
+                rightMargin: units.dp(16)
+                topMargin: units.dp(16)
                 bottomMargin: hasActions ? 0 : units.dp(16)
             }
-            id: placeholder
         }
 
         Item {
@@ -140,14 +136,16 @@ Popover {
 
                 text: negativeButtonText
                 textColor: Theme.accentColor
+
                 anchors {
                     top: parent.top
-                    topMargin: units.dp(8)
                     right: positiveButton.left
-                    rightMargin: units.dp(8)
                     bottom: parent.bottom
+                    topMargin: units.dp(8)
+                    rightMargin: units.dp(8)
                     bottomMargin: units.dp(8)
                 }
+
                 onClicked: {
                     close();
                     rejected();
@@ -159,14 +157,16 @@ Popover {
 
                 text: positiveButtonText
                 textColor: Theme.accentColor
+
                 anchors {
                     top: parent.top
                     topMargin: units.dp(8)
-                    right: parent.right
                     rightMargin: units.dp(16)
-                    bottom: parent.bottom
                     bottomMargin: units.dp(8)
+                    right: parent.right
+                    bottom: parent.bottom
                 }
+
                 onClicked: {
                     close()
                     accepted();
@@ -175,7 +175,4 @@ Popover {
         }
     }
 
-    function show() {
-        open(parent)
-    }
 }