+++ CHANGELOG.md
@@ -6,6 +6,7 @@ QML Material CHANGELOG
  * Rename list item and IconButton onTriggered to onClicked
  * [Toolbar] Add support for custom and extended content in the toolbar
  * [PageStack] Make Page a FocusScope
+ * [Dialog Refactor the dialog code to use anchors
  * Minor bug fixes and code improvements